DWPF-1767 Improve new_finance_year command (#505)
Whilst trying to run the `new_finance_year` command, we ran into some
bad data that had made it's way into the database. What we found was a
used financial code with a NAC that had an invalid economic budget code.
To resolve the issue we fixed the data by changing the NAC's economic
budget code.

This commit extends the `new_finance_year` command to detect this issue
before changing the financial year. It also serves to document the issue
we faced.