FFT-95 Optimise edit forecast database queries (#546)
I set out to improve the performance of the edit forecast page. On my
local machine it took over 6 seconds just to fetch the initial HTML.

Our local dataset contains over 400 rows in the table on that page, so I
suspected we had an n+1 database query problem. After some digging it
turns out I was right and there were over 17,000 queries!

After a couple of days spent optimising code, mostly through the use of
`select_related` and `prefetch_related`, I've managed to get the number
of queries down to 13 and the page load time down to around 300ms!

Here is a brief list of the optimisations:
- apply `select_related` and `prefetch_related` where needed
- store the `current_financial_year` on the `request` object
- apply `cached_property` to view methods being used in the template
- pass the `current_financial_year` as context into the financial code
serializer
- add budget amount annotation to queryset
- small javascript refactor to avoid needless object creation

There was a more ambitious optimising around the
`current_financial_year` using a `ContextVar`, but I chose to leave it
out and go with storing it on the request object as it's simpler.

The react code is now the bottleneck on that page :)