Revert "Debugpy configuration setup (#581)" (#586)
This reverts commit 2837b7e30cee0067d8fb6bd40d11b90555f5e1b7.

I'm reverting this change for a few of reasons. I'm not keen on having a
lot of debug code in `manage.py` as I consider it core production code.
The `debugpy` package was included into the production python packages.
And, I believe there is a better approach where we wrap the `runserver`
command instead.

I would like to take more time to explore other options before
committing to this change.