FFT-318 Fix orphaned payroll forecasts (#717)
Orphaned payroll forecasts would occur when a financial code is no being driven
by the payroll forecast. This would usually happen when a vacancy was created
with one programme code and then switched to another. If the first programme
code was no longer used by any employee or vacancy in that cost centre, the
forecast for that financial code would be left, causing the numbers to be wrong.

The solution is to remove any forecasts which are orphaned when the payroll
forecast of a cost centre is updated.